Add "language features" to the C++ parser
Currently the parser is a pure C++ parser with C++11 elements, without
any option to disable C++11 or to make it more C compatible. C
compatibility in particular is a big challenge since many C++ tokens can
be used as identifiers in C, and C also has some constructs that are not
valid in C++ (like designated initializers or casted braced init lists).

This patch adds "language features" which are flags that enable specific
language support in the parser. Currently flags exist for C99, C++ and
C++11. Code that invokes the parser can use the new
setLanguageFeatures() method to change the current set of supported
features (the default is C++ | C++11). This is useful for instance for
C-specific projects.

CppLanguageSupport::createParseJob has been modified to check for a
"Language" entry in the configuration of the project the parsed file
belongs to. Project controllers can set this option if they know that
the project is limited to C, C++ or C++11 files. Other means (for
instance, based on file extension of Makefile analysis) could be
implemented in the future.

Conflicts:
	languages/cpp/cpplanguagesupport.cpp
	languages/cpp/preprocessjob.cpp